+++ packages/typespec-rust/test/cadlranch/server/endpoint/not-defined/tests/not_defined_client_test.rs
@@ -1,0 +1,11 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+//
+// Licensed under the MIT License. See License.txt in the project root for license information.
+
+use cadl_noendpoint::NotDefinedClient;
+
+#[async_std::test]
+async fn valid() {
+    let client = NotDefinedClient::with_no_credential("http://localhost:3000", None).unwrap();
+    client.valid(None).await.unwrap();
+}